+++ src/main/java/org/bytedeco/javacv/FFmpegFrameRecorder.java
@@ -71,7 +71,6 @@ import org.bytedeco.javacpp.IntPointer;
 import org.bytedeco.javacpp.Loader;
 import org.bytedeco.javacpp.Pointer;
 import org.bytedeco.javacpp.PointerPointer;
-import org.bytedeco.javacpp.PointerScope;
 import org.bytedeco.javacpp.ShortPointer;
 
 import org.bytedeco.ffmpeg.avcodec.*;
@@ -319,13 +318,7 @@ public class FFmpegFrameRecorder extends FrameRecorder {
         }
     }
 
-    static WriteCallback writeCallback = new WriteCallback();
-    static {
-        PointerScope s = PointerScope.getInnerScope();
-        if (s != null) {
-            s.detach(writeCallback);
-        }
-    }
+    static WriteCallback writeCallback = new WriteCallback().retainReference();
 
     private OutputStream outputStream;
     private boolean closeOutputStream;